refactored RideIntegrationTests and added comments
