+++ src/test/java/com/bussin/SpringBack/rideTests/RideCreateIntegrationTests.java
@@ -2,26 +2,15 @@ package com.bussin.SpringBack.rideTests;
 
 import com.bussin.SpringBack.TestObjects;
 import com.bussin.SpringBack.integrationTestAuth.CognitoLogin;
-import com.bussin.SpringBack.models.DriverDTO;
-import com.bussin.SpringBack.models.PlannedRoute;
-import com.bussin.SpringBack.models.PlannedRouteDTO;
-import com.bussin.SpringBack.models.Ride;
-import com.bussin.SpringBack.models.RideCreationDTO;
-import com.bussin.SpringBack.models.RideDTO;
-import com.bussin.SpringBack.models.User;
-import com.bussin.SpringBack.models.UserDTO;
+import com.bussin.SpringBack.models.*;
 import com.bussin.SpringBack.services.DriverService;
 import com.bussin.SpringBack.services.PlannedRouteService;
 import com.bussin.SpringBack.services.RideService;
 import com.bussin.SpringBack.services.UserService;
 import com.bussin.SpringBack.testConfig.H2JpaConfig;
 import com.bussin.SpringBack.testConfig.TestContextConfig;
-import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.hc.client5.http.classic.methods.HttpDelete;
-import org.apache.hc.client5.http.classic.methods.HttpGet;
 import org.apache.hc.client5.http.classic.methods.HttpPost;
-import org.apache.hc.client5.http.classic.methods.HttpPut;
 import org.apache.hc.client5.http.classic.methods.HttpUriRequest;
 import org.apache.hc.client5.http.impl.classic.CloseableHttpResponse;
 import org.apache.hc.client5.http.impl.classic.HttpClientBuilder;
@@ -40,7 +29,6 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.io.IOException;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 
@@ -51,7 +39,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public class RideIntegrationTests {
+public class RideCreateIntegrationTests {
     @LocalServerPort
     private int port;
 
@@ -82,116 +70,20 @@ public class RideIntegrationTests {
 
     private static final String AUTHORIZATION_HEADER = "Authorization";
 
+    /**
+     * Authenticate JWTToken and create a new TestObject user before each tests
+     * @throws IOException
+     */
     @BeforeEach
     private void setUp() throws IOException {
         idToken = "Bearer " + cognitoLogin.getAuthToken();
         userService.createNewUser(TestObjects.COGNITO_USER_DTO);
     }
 
-    @Test
-    public void getAllRides_noRides_success() throws IOException {
-        HttpUriRequest request = new HttpGet(baseUrl + port + "/api/v1" +
-                "/ride");
-        request.setHeader(AUTHORIZATION_HEADER, idToken);
-
-        CloseableHttpResponse httpResponse =
-                HttpClientBuilder.create().build().execute(request);
-
-        assertEquals(httpResponse.getCode(), 200);
-    }
-
-    @Test
-    public void getAllRides_success() throws IOException {
-        UserDTO userDTO = TestObjects.USER_DTO.clone();
-
-        User user = userService.createNewUser(userDTO);
-
-        DriverDTO driverDTO = TestObjects.DRIVER_DTO.clone();
-
-        driverService.addNewDriver(user.getId(), driverDTO);
-
-        PlannedRouteDTO plannedRouteDTO = TestObjects.PLANNED_ROUTE_DTO.clone();
-
-        PlannedRoute plannedRoute =
-                plannedRouteService.createNewPlannedRoute(plannedRouteDTO,
-                        driverDTO.getCarPlate());
-
-        RideDTO rideDTO = TestObjects.RIDE_DTO.clone();
-
-        Ride ride = rideService.createNewRide(rideDTO, user.getId(),
-                plannedRoute.getId());
-
-        HttpUriRequest request = new HttpGet(baseUrl + port + "/api/v1" +
-                "/ride");
-        request.setHeader(AUTHORIZATION_HEADER, idToken);
-
-        CloseableHttpResponse httpResponse =
-                HttpClientBuilder.create().build().execute(request);
-
-        List<Ride> rides =
-                objectMapper.readValue(httpResponse.getEntity().getContent(),
-                        new TypeReference<>() {
-                        });
-
-        rideDTO.setId(ride.getId());
-
-        RideDTO dest = RideDTO.builder().build();
-        modelMapper.map(rides.get(0), dest);
-        assertEquals(dest, rideDTO);
-    }
-
-    @Test
-    public void getRideByID_success() throws IOException {
-        UserDTO userDTO = TestObjects.USER_DTO.clone();
-        userDTO.setIsDriver(true);
-
-        User user = userService.createNewUser(userDTO);
-
-        DriverDTO driverDTO = TestObjects.DRIVER_DTO.clone();
-
-        driverService.addNewDriver(user.getId(), driverDTO);
-
-        PlannedRouteDTO plannedRouteDTO = TestObjects.PLANNED_ROUTE_DTO.clone();
-
-        PlannedRoute plannedRoute =
-                plannedRouteService.createNewPlannedRoute(plannedRouteDTO,
-                        driverDTO.getCarPlate());
-
-        RideDTO rideDTO = TestObjects.RIDE_DTO.clone();
-
-        Ride ride = rideService.createNewRide(rideDTO, user.getId(),
-                plannedRoute.getId());
-
-        HttpUriRequest request = new HttpGet(baseUrl + port + "/api/v1" +
-                "/ride/" + ride.getId());
-        request.setHeader(AUTHORIZATION_HEADER, idToken);
-
-        CloseableHttpResponse httpResponse =
-                HttpClientBuilder.create().build().execute(request);
-
-        Ride rideResult =
-                objectMapper.readValue(httpResponse.getEntity().getContent(),
-                        Ride.class);
-
-        rideDTO.setId(ride.getId());
-
-        RideDTO dest = RideDTO.builder().build();
-        modelMapper.map(rideResult, dest);
-        assertEquals(dest, rideDTO);
-    }
-
-    @Test
-    public void getRideByID_noRide_404() throws IOException {
-        HttpUriRequest request = new HttpGet(baseUrl + port + "/api/v1" +
-                "/ride/" + UUID.randomUUID());
-        request.setHeader(AUTHORIZATION_HEADER, idToken);
-
-        CloseableHttpResponse httpResponse =
-                HttpClientBuilder.create().build().execute(request);
-
-        assertEquals(404, httpResponse.getCode());
-    }
-
+    /**
+     * Create a new ride with valid credentials and parameters success
+     * @throws IOException
+     */
     @Test
     public void createNewRide_success() throws IOException {
         UserDTO userDTO = TestObjects.USER_DTO.clone();
@@ -242,6 +134,10 @@ public class RideIntegrationTests {
         assertEquals(dest, rideDTO);
     }
 
+    /**
+     * Create a new ride with invalid parameter throws 400 BAD_REQUEST
+     * @throws IOException
+     */
     @Test
     public void createNewRide_invalidParams_400() throws IOException {
         UserDTO userDTO = TestObjects.USER_DTO.clone();
@@ -285,6 +181,10 @@ public class RideIntegrationTests {
         assertEquals(400, httpResponse.getCode());
     }
 
+    /**
+     * Create a new ride with passengers over capacity throws 400 BAD_REQUEST
+     * @throws IOException
+     */
     @Test
     public void createNewRide_passengersOverCapacity_400() throws IOException {
         UserDTO userDTO = TestObjects.USER_DTO.clone();
@@ -328,6 +228,11 @@ public class RideIntegrationTests {
         assertEquals(400, httpResponse.getCode());
     }
 
+    /**
+     * Create a new ride on top of the current rides,
+     * when added capacity more than the capacity will throw 400 BAD_REQUEST
+     * @throws IOException
+     */
     @Test
     public void createNewRide_multipleRidesOverCapacity_400() throws IOException {
         UserDTO userDTO = TestObjects.USER_DTO.clone();
@@ -378,6 +283,10 @@ public class RideIntegrationTests {
         assertEquals(400, httpResponse.getCode());
     }
 
+    /**
+     * Create a new ride with no user found throws 404 NOT_FOUND
+     * @throws IOException
+     */
     @Test
     public void createNewRide_missingUser_404() throws IOException {
         UserDTO userDTO = TestObjects.USER_DTO.clone();
@@ -420,6 +329,10 @@ public class RideIntegrationTests {
         assertEquals(404, httpResponse.getCode());
     }
 
+    /**
+     * Create a new ride with no route found throws 404 NOT_FOUND
+     * @throws IOException
+     */
     @Test
     public void createNewRide_noRoute_404() throws IOException {
         UserDTO userDTO = TestObjects.USER_DTO.clone();
@@ -454,189 +367,6 @@ public class RideIntegrationTests {
         request.setEntity(entity);
         request.setHeader("Accept", "application/json");
         request.setHeader("Content-type", "application/json");
-
-        CloseableHttpResponse httpResponse =
-                HttpClientBuilder.create().build().execute(request);
-
-        assertEquals(404, httpResponse.getCode());
-    }
-
-    @Test
-    public void updateRideById_success() throws IOException {
-        UserDTO userDTO = TestObjects.USER_DTO.clone();
-        userDTO.setIsDriver(true);
-
-        User user = userService.createNewUser(userDTO);
-
-        DriverDTO driverDTO = TestObjects.DRIVER_DTO.clone();
-
-        driverService.addNewDriver(user.getId(), driverDTO);
-
-        PlannedRouteDTO plannedRouteDTO = TestObjects.PLANNED_ROUTE_DTO.clone();
-
-        PlannedRoute plannedRoute =
-                plannedRouteService.createNewPlannedRoute(plannedRouteDTO,
-                        driverDTO.getCarPlate());
-
-        RideDTO rideDTO = TestObjects.RIDE_DTO.clone();
-
-        RideDTO updatedRideDTO = TestObjects.RIDE_DTO.clone();
-        updatedRideDTO.setPassengers(2);
-
-        Ride ride = rideService.createNewRide(rideDTO, user.getId(),
-                plannedRoute.getId());
-
-        HttpUriRequest request = new HttpPut(baseUrl + port + "/api/v1" +
-                "/ride/" + ride.getId());
-        request.setHeader(AUTHORIZATION_HEADER, idToken);
-
-        StringEntity entity = new StringEntity(objectMapper
-                .writeValueAsString(updatedRideDTO));
-
-        request.setEntity(entity);
-        request.setHeader("Accept", "application/json");
-        request.setHeader("Content-type", "application/json");
-
-        CloseableHttpResponse httpResponse =
-                HttpClientBuilder.create().build().execute(request);
-
-        Ride rideResult =
-                objectMapper.readValue(httpResponse.getEntity().getContent(),
-                        Ride.class);
-
-        updatedRideDTO.setId(ride.getId());
-
-        RideDTO dest = RideDTO.builder().build();
-        modelMapper.map(rideResult, dest);
-        assertEquals(dest, updatedRideDTO);
-    }
-
-    @Test
-    public void updateRideById_invalidParams_400() throws IOException {
-        UserDTO userDTO = TestObjects.USER_DTO.clone();
-        userDTO.setIsDriver(true);
-
-        User user = userService.createNewUser(userDTO);
-
-        DriverDTO driverDTO = TestObjects.DRIVER_DTO.clone();
-
-        driverService.addNewDriver(user.getId(), driverDTO);
-
-        PlannedRouteDTO plannedRouteDTO = TestObjects.PLANNED_ROUTE_DTO.clone();
-
-        PlannedRoute plannedRoute =
-                plannedRouteService.createNewPlannedRoute(plannedRouteDTO,
-                        driverDTO.getCarPlate());
-
-        RideDTO rideDTO = TestObjects.RIDE_DTO.clone();
-
-        RideDTO updatedRideDTO = RideDTO.builder()
-                .passengers(1000)
-                .timestamp(rideDTO.getTimestamp())
-                .build();
-
-        Ride ride = rideService.createNewRide(rideDTO, user.getId(),
-                plannedRoute.getId());
-
-        HttpUriRequest request = new HttpPut(baseUrl + port + "/api/v1" +
-                "/ride/" + ride.getId());
-        request.setHeader(AUTHORIZATION_HEADER, idToken);
-
-        StringEntity entity = new StringEntity(objectMapper
-                .writeValueAsString(updatedRideDTO));
-
-        request.setEntity(entity);
-        request.setHeader("Accept", "application/json");
-        request.setHeader("Content-type", "application/json");
-
-        CloseableHttpResponse httpResponse =
-                HttpClientBuilder.create().build().execute(request);
-
-        assertEquals(400, httpResponse.getCode());
-    }
-
-    @Test
-    public void updateRideById_noRide_404() throws IOException {
-        UserDTO userDTO = TestObjects.USER_DTO.clone();
-
-        User user = userService.createNewUser(userDTO);
-
-        DriverDTO driverDTO = TestObjects.DRIVER_DTO.clone();
-
-        driverService.addNewDriver(user.getId(), driverDTO);
-
-        PlannedRouteDTO plannedRouteDTO = TestObjects.PLANNED_ROUTE_DTO.clone();
-        plannedRouteDTO.setCapacity(3);
-
-        PlannedRoute plannedRoute =
-                plannedRouteService.createNewPlannedRoute(plannedRouteDTO,
-                        driverDTO.getCarPlate());
-
-        RideDTO updatedRideDTO = TestObjects.RIDE_DTO.clone();
-        updatedRideDTO.setPassengers(3);
-
-        HttpUriRequest request = new HttpPut(baseUrl + port + "/api/v1" +
-                "/ride/" + plannedRoute.getId());
-        request.setHeader(AUTHORIZATION_HEADER, idToken);
-
-        StringEntity entity = new StringEntity(objectMapper
-                .writeValueAsString(updatedRideDTO));
-
-        request.setEntity(entity);
-        request.setHeader("Accept", "application/json");
-        request.setHeader("Content-type", "application/json");
-
-        CloseableHttpResponse httpResponse =
-                HttpClientBuilder.create().build().execute(request);
-
-        assertEquals(404, httpResponse.getCode());
-    }
-
-    @Test
-    public void deleteRideByID_success() throws IOException {
-        UserDTO userDTO = TestObjects.USER_DTO.clone();
-        userDTO.setIsDriver(true);
-
-        User user = userService.createNewUser(userDTO);
-
-        DriverDTO driverDTO = TestObjects.DRIVER_DTO.clone();
-
-        driverService.addNewDriver(user.getId(), driverDTO);
-
-        PlannedRouteDTO plannedRouteDTO = TestObjects.PLANNED_ROUTE_DTO;
-
-        PlannedRoute plannedRoute =
-                plannedRouteService.createNewPlannedRoute(plannedRouteDTO,
-                        driverDTO.getCarPlate());
-
-        RideDTO rideDTO = TestObjects.RIDE_DTO.clone();
-
-        Ride ride = rideService.createNewRide(rideDTO, user.getId(),
-                plannedRoute.getId());
-
-        HttpUriRequest request = new HttpDelete(baseUrl + port + "/api/v1" +
-                "/ride/" + ride.getId());
-        request.setHeader(AUTHORIZATION_HEADER, idToken);
-
-        CloseableHttpResponse httpResponse =
-                HttpClientBuilder.create().build().execute(request);
-
-        Ride rideResult =
-                objectMapper.readValue(httpResponse.getEntity().getContent(),
-                        Ride.class);
-
-        rideDTO.setId(ride.getId());
-
-        RideDTO dest = RideDTO.builder().build();
-        modelMapper.map(rideResult, dest);
-        assertEquals(dest, rideDTO);
-    }
-
-    @Test
-    public void deleteRideByID_noRide_404() throws IOException {
-        HttpUriRequest request = new HttpDelete(baseUrl + port + "/api/v1" +
-                "/ride/" + UUID.randomUUID());
-        request.setHeader(AUTHORIZATION_HEADER, idToken);
 
         CloseableHttpResponse httpResponse =
                 HttpClientBuilder.create().build().execute(request);
